fix a weird bug
in which the last used community for the current user was being shown
as the community for the project being viewed. i’m still not 100% sure
why, but this fixes it. (merging the inline-edit template back into the
inlinePostInput directive wasn’t part of the fix, it was just to reduce
the number of different files involved in this tortuous chain)